+++ packages/themed/CHANGELOG.md
@@ -3,6 +3,45 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 4.0.0-rc.2 (2022-04-27)
+
+
+### Bug Fixes
+
+* **SearchBar:** typescript errors ([#3437](https://github.com/react-native-elements/react-native-elements/issues/3437)) ([ec01b65](https://github.com/react-native-elements/react-native-elements/commit/ec01b65cbbb3a3795e0450a4c744acffa2b31d56))
+* **SpeedDial.Action:** wrap with pressable ([#3480](https://github.com/react-native-elements/react-native-elements/issues/3480)) ([6de7aa0](https://github.com/react-native-elements/react-native-elements/commit/6de7aa0985edc90a0325244227eda9e4aab71b60))
+* **SpeedDial:** Left placement for Actions ([#3479](https://github.com/react-native-elements/react-native-elements/issues/3479)) ([d006fe9](https://github.com/react-native-elements/react-native-elements/commit/d006fe9f532430a00931877be8a5cda4b4027942))
+* **ThemeProvider:** add `children` prop explicitly ([#3477](https://github.com/react-native-elements/react-native-elements/issues/3477)) ([45f9523](https://github.com/react-native-elements/react-native-elements/commit/45f9523300d9bcc74837ddc01dc6b787cb4be54b))
+* **Theming:** extend default theme ([#3450](https://github.com/react-native-elements/react-native-elements/issues/3450)) ([587d8db](https://github.com/react-native-elements/react-native-elements/commit/587d8dbf8aef0beb9cd42683f16e52bbbe3b99c1))
+
+
+
+# 4.0.0-rc.1 (2022-03-28)
+
+
+### Bug Fixes
+
+* pkg entry point for react-native cli ([#3403](https://github.com/react-native-elements/react-native-elements/issues/3403)) ([2c4bf8a](https://github.com/react-native-elements/react-native-elements/commit/2c4bf8a4a31498ce4dd2aefe22cff5cfc13363f1))
+* remove Partial Color Type ([#3418](https://github.com/react-native-elements/react-native-elements/issues/3418)) ([c9c2bdc](https://github.com/react-native-elements/react-native-elements/commit/c9c2bdc58198392727c1b6affccf6201724a2391))
+
+
+
+# 4.0.0-rc.0 (2022-03-16)
+
+
+### Bug Fixes
+
+* collectCoverage ([483af88](https://github.com/react-native-elements/react-native-elements/commit/483af88803a48f343ff2965d9770f9a5ea71a033))
+
+
+### Features
+
+* **Skeleton:** new universal component package ([#3357](https://github.com/react-native-elements/react-native-elements/issues/3357)) ([1cca2db](https://github.com/react-native-elements/react-native-elements/commit/1cca2db1967330ef42a67c9b885a460ed27a4762))
+
+
+
+
+
 # 4.0.0-rc.1 (2022-03-28)